api/generate-asertions/requirements.txt
fastapi==0.101.1
uvicorn[standard]==0.23.2
pydantic==2.7.1
requests==2.32.0
aiokafka==0.12.0
python-dotenv==1.0.0
